requirements.txt
python-dotenv==1.0.0
aiortc==1.5.0
websockets==13.0.0
numpy
av
opencv-python
pyaudio
